### 2024-06-06 #6852 @editorialbot
[PRE REVIEW]: GluPredKit: A Python Package for Blood Glucose Prediction and Evaluation
**Submitting author:** @miriamkw (<a href="http://orcid.org/0009-0000-6919-8164">Miriam Kopperstad Wolff</a>)
**Repository:** https://github.com/miriamkw/GluPredKit
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
