### 2024-06-07 #6852 @miriamkw
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/9402969762?check_suite_focus=true).

My sincere apologies. I mistakenly inserted the link to the branch in the input field for the branch name with `paper.md`, instead of just writing the branch name. I hope this can be corrected. Please let me know what steps I should take to fix this. Thank you.
