### 2024-06-09 #6852 @Kevin-Mattheus-Moerman
@miriamkw thanks for this submission. I am the AEiC in this track and here to help with initial steps.

For the moment can you work on these points for your paper:
- [ ] In your affiliations, please spell out USA as United States of America
- [ ] It looks like a word may be missing in the sentence starting "The GluPredKit workflow is typically through the Command Line Interface..."
- [ ] Can you check if you can address any of the above potentially missing DOIs? It is okay if some of them truly do not have a DOI. Thanks
