### 2024-06-09 efd3b60
Removed unused reference
