+++ paper.bib
@@ -93,15 +93,6 @@ SIGNIFICANCE: These consensus guidelines are designed to improve performance and
 
 
 
-@unpublished{Oh2024,
-  title     = {GluPredKit – An Open-Source Researcher Platform for Blood Glucose Prediction},
-  author    = {Miriam Kopperstad Wolff and Sam Royston and Anders Lyngvi Fougner and Martin Steinert and Hans Georg Schaathun and Rune Volden},
-  note      = {Under review},
-  year      = {2024}
-}
-
-
-
 
 @unpublished{oh2024thesis,
   title     = {Improving Blood Glucose Prediction for People with T1DM During Physical Activity Using Machine Learning on Participant Collected Data},
