### 2024-06-09 5c36f25
Fixed correct abbreviations for command line interface
