### 2024-06-09 #6852 @miriamkw
Hi @Kevin-Mattheus-Moerman, thank you for your feedback. I appreciate your assistance in improving our submission. I have made the updates in response to your comments.

The references with missing DOIs do not have DOIs. Additionally, some of the papers are still in the process of being published. I have marked these papers as @unpublished in the paper.bib file to reflect their current status.

Please let me know if there are any further changes or additional details needed.

Best regards, Miriam
