### 2024-06-09 #6852 @Kevin-Mattheus-Moerman
@miriamkw thanks for making those changes. You cannot cite unpublished works. We recommend that you post preprints for these instead (assuming they are your work) which are citable.
