### 2024-06-13 #6852 @miriamkw
> @miriamkw okay, thanks for the update. Have you checked the other points I mentioned?

Yes, thank you. The suggested changes, excluding those about the missing DOIs, have been addressed in the version I generated four days ago.
