### 2024-06-14 #6852 @jromanowska
Hi, @miriamkw  :wave:
I will take care of the process from now on. I will go through the paper and look at the structure of the repository. Then I'll try to find reviewers, but please let me know if you have any suggestions (or oppositions). When naming reviewers, write their GitHub handles _without_ `@`, so that they _won't_ get notified.
