+++ paper.bib
@@ -84,25 +84,30 @@ SIGNIFICANCE: These consensus guidelines are designed to improve performance and
 
 
 
-@unpublished{Wolff2024,
-  title     = {GluPredKit – An Open-Source Researcher Platform for Blood Glucose Prediction},
-  author    = {Miriam Kopperstad Wolff and Sam Royston and Anders Lyngvi Fougner and Martin Steinert and Hans Georg Schaathun and Rune Volden},
-  note      = {Under review},
-  year      = {2024}
+@misc{Wolff2024,
+      title={GluPredKit: Development and User Evaluation of a Standardization Software for Blood Glucose Prediction},
+      author={Miriam K. Wolff and Sam Royston and Anders Lyngvi Fougner and Hans Georg Schaathun and Martin Steinert and Rune Volden},
+      year={2024},
+      eprint={2406.08915},
+      archivePrefix={arXiv},
+      primaryClass={id='cs.SE' full_name='Software Engineering' is_active=True alt_name=None in_archive='cs' is_general=False description='Covers design tools, software metrics, testing and debugging, programming environments, etc. Roughly includes material in all of ACM Subject Classes D.2, except that D.2.4 (program verification) should probably have Logics in Computer Science as the primary subject area.'}
+}
+
+
+@mastersthesis{oh2024thesis,
+  title={Improving Blood Glucose Prediction for People with T1DM During Physical Activity Using Machine Learning on Participant Collected Data},
+  author={Oh, Doyoung},
+  school={UiT The Arctic University of Norway},
+  year={2024},
+  type={Mastergradsoppgave},
+  url={https://hdl.handle.net/10037/33803}
 }
 
 
 
 
-@unpublished{oh2024thesis,
-  title     = {Improving Blood Glucose Prediction for People with T1DM During Physical Activity Using Machine Learning on Participant Collected Data},
-  author    = {Doyoung Oh},
-  note      = {Unpublished Master's thesis, Faculty of Science and Technology, Department of Computer Science, UiT The Arctic University of Norway},
-  year      = {2024},
-  month     = {May},
-  type      = {INF-3990-1 Master’s Thesis in Computer Science},
-  supervisor= {Professor Eirik Årsand [UiT/NSE], Associate Professor André Henriksen [UiT], PhD candidate Miriam Wolff [NTNU], Dr. Phuong Dinh Ngo [NSE]}
-}
+
+
 
 
 @inproceedings{marling2020ohiot1dm,
