+++ paper.bib
@@ -99,7 +99,6 @@ SIGNIFICANCE: These consensus guidelines are designed to improve performance and
   author={Oh, Doyoung},
   school={UiT The Arctic University of Norway},
   year={2024},
-  type={Mastergradsoppgave},
   url={https://hdl.handle.net/10037/33803}
 }
 
