### 2024-06-19 #6852 @mikeaalv
This looks interesting. However, I'm overwhelmed by dues recently and have to decline. Thanks for the invitation.
