### 2024-06-20 #6852 @JonnyTran
> :wave: @mikeaalv & @JonnyTran, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @jromanowska, I'd love to review the package. I've reviewed 3 packages before, so quite familiar with the guidelines.
