### 2024-06-20 #6904 @editorialbot
[REVIEW]: GluPredKit: A Python Package for Blood Glucose Prediction and Evaluation
**Submitting author:** @miriamkw (<a href="http://orcid.org/0009-0000-6919-8164">Miriam Kopperstad Wolff</a>)
**Repository:** https://github.com/miriamkw/GluPredKit
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.12
