### 2024-06-20 #6904 @jromanowska
@miriamkw , I see that the references that were marked by editorialbot as having no DOIs are fine as these are not published with a DOI, but maybe you can double-check that and confirm here?
