### 2024-06-20 #6904 @miriamkw
@jromanowska I confirm that they do not have a DOI. Two are master's theses, one is a preprint of my own work, and the last is a paper from a workshop.
