### 2024-06-21 a4ed446
Updated open aps parser to parse zip files directly
