+++ glupredkit/parsers/open_aps.py
@@ -1,8 +1,11 @@
-"""
-The Open APS parser is processing the .pkl file produced by using the data cleaner by Harry Emerson: https://github.com/hemerson1/OpenAPS_Cleaner.
-"""
 import pandas as pd
-import pickle
+import os
+import zipfile
+import re
+import numpy as np
+import psutil
+import json
+from datetime import datetime
 from .base_parser import BaseParser
 
 
@@ -12,20 +15,410 @@ class Parser(BaseParser):
 
     def __call__(self, file_path: str, *args):
         """
-        file_path -- the file path to the processed data including <file_name>.pkl.
+        file_path -- the file path to the folder that contains all the .zip files of the OpenAPS data.
         """
-        df = pd.read_csv(file_path, low_memory=False)
-        relevant_columns = ['date', 'bg', 'basal', 'carbs', 'bolus', 'PtID']
-        df = df[relevant_columns]
-        df = df.rename(columns={'bg': 'CGM', 'PtID': 'id'})
-        df['date'] = pd.to_datetime(df['date'])
+        merged_df = pd.DataFrame()
 
-        # Sort by 'id' and the index (date)
-        df = df.sort_values(by=['id', 'date'])
+        # List all files in the folder
+        files = [el for el in os.listdir(file_path) if el.endswith('.zip')]  # 142 subjects
 
-        # Set date to index
-        df = df.set_index('date')
+        # Process the files one by one
+        for file in files:
+            if file == 'AndroidAPS Uploader.zip':
+                with zipfile.ZipFile(file_path + 'AndroidAPS Uploader.zip', 'r') as zip_ref:
+                    # Find unique ids
+                    all_ids = np.unique([file.split('/')[0] for file in zip_ref.namelist() if not file.split('/')[0] == ''])
 
+                    for subject_id in all_ids:
+                        print(f'Processing {subject_id}...')
+
+                        def get_relevant_files(name):
+                            relevant_files = []
+
+                            for file_name in zip_ref.namelist():
+                                if file_name.startswith(subject_id) and file_name.endswith(f'{name}.json'):
+                                    relevant_files.append(file_name)
+                                    # print("FILE NAME", file_name)
+
+                            # Check whether the file is found
+                            if not relevant_files:
+                                print(f"No files found  containing '{name}' in the name!")
+
+                            return relevant_files
+
+                        # Blood glucose
+                        entries_files = get_relevant_files('BgReadings')
+
+                        # Carbohydrates and insulin
+                        treatments_files = get_relevant_files('Treatments')
+
+                        # Basal rates
+                        basal_files = get_relevant_files('APSData')
+
+                        # Temporary basal rates
+                        temp_basal_files = get_relevant_files('TemporaryBasals')
+
+                        # Skip to next iteration if entries_files is empty
+                        if not entries_files or not treatments_files or not basal_files:
+                            print("Skipping to next subject...")
+                            continue
+
+                        all_entries_dfs = []
+                        for entries_file in entries_files:
+                            with zip_ref.open(entries_file) as f:
+                                entries_df = pd.read_json(f, convert_dates=False)
+
+                            entries_df['date'] = pd.to_datetime(entries_df['date'], unit='ms')
+                            entries_df['value'] = pd.to_numeric(entries_df['value'])
+                            entries_df = entries_df[['date', 'value']]
+                            entries_df.rename(columns={'value': 'CGM'}, inplace=True)
+                            entries_df.set_index('date', inplace=True)
+                            entries_df.sort_index(inplace=True)
+                            all_entries_dfs.append(entries_df)
+
+                        df = pd.concat(all_entries_dfs)
+                        df = df.resample('5min').mean()
+
+                        carbs_dfs = []
[…]
