+++ glupredkit/parsers/ohio_t1dm.py
@@ -110,6 +110,8 @@ class Parser(BaseParser):
         df = pd.merge(df, df_basal, on="date", how='outer')
         df['basal'] = df['basal'].ffill()
 
+        df['insulin'] = df['bolus'] + df['basal'] * 5 / 60
+
         # Heart rate
         df = merge_data_type_into_dataframe(df, dataframes, 'basis_heart_rate', 'heartrate')
 
