+++ requirements.txt
@@ -6,6 +6,7 @@ pandas
 scikit-learn
 tidepool-data-science-project
 numpy
+psutil
 aiohttp
 click
 dill
