+++ glupredkit/parsers/open_aps.py
@@ -180,242 +180,240 @@ class Parser(BaseParser):
                         # Add id to a column
                         df['id'] = subject_id
 
-                        # We save the merged dataframe as csv, but update it for each iteration in case something goes wrong
                         merged_df = pd.concat([df, merged_df], ignore_index=False)
-
                         print(f"Current memory usage: {get_memory_usage()} MB")
 
-            id_name = file.split('.')[0]
+            else:
+                id_name = file.split('.')[0]
 
-            print(f'Processing {file}...')
-            zip_file_path = file_path + file
-            sub_folder_path = 'direct-sharing-31/'
+                print(f'Processing {file}...')
+                zip_file_path = file_path + file
+                sub_folder_path = 'direct-sharing-31/'
 
-            # Open the zip file
-            with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
-                def get_relevant_files(name):
-                    relevant_files = []
+                # Open the zip file
+                with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
+                    def get_relevant_files(name):
+                        relevant_files = []
 
-                    for file_name in zip_ref.namelist():
-                        if len(file_name.split('/')) == 2:
-                            if name in file_name.lower():
-                                if file_name.endswith('.json'):
-                                    relevant_files.append(file_name)
+                        for file_name in zip_ref.namelist():
+                            if len(file_name.split('/')) == 2:
+                                if name in file_name.lower():
+                                    if file_name.endswith('.json'):
+                                        relevant_files.append(file_name)
 
-                    # Check whether the file is found
-                    if not relevant_files:
-                        print(f"No files found  containing '{name}' in the name!")
+                        # Check whether the file is found
+                        if not relevant_files:
+                            print(f"No files found  containing '{name}' in the name!")
 
-                    return relevant_files
+                        return relevant_files
 
-                # Blood glucose
-                entries_files = get_relevant_files('entries')
-                all_entries_dfs = []
-                for entries_file in entries_files:
-                    with zip_ref.open(entries_file) as f:
-                        if entries_file.startswith(sub_folder_path):
-                            try:
-                                entries_df = pd.read_json(f, convert_dates=False)
-                            except ValueError as e:
-                                # Reset the file handle to the start
-                                f.seek(0)
-                                entries_df = pd.DataFrame()
-                                print(f"Error parsing JSON directly: {e}")
-                                for line in f:
-                                    try:
-                                        data = json.loads(line)
-                                        entries_df = entries_df._append(data, ignore_index=True)
-                                    except json.JSONDecodeError as json_err:
-                                        print(f"Skipping line due to error: {json_err}")
-                        else:
-                            entries_df = pd.read_json(f, convert_dates=False, lines=True)
-                        if entries_df.empty:
-                            continue
-                        entries_df['dateString'] = entries_df['dateString'].apply(parse_datetime_without_timezone)
-                        entries_df['sgv'] = pd.to_numeric(entries_df['sgv'])
-                        entries_df = entries_df[['dateString', 'sgv']]
-                        entries_df.rename(columns={'sgv': 'CGM', 'dateString': 'date'}, inplace=True)
-                        entries_df.set_index('date', inplace=True)
-                        entries_df.sort_index(inplace=True)
-                        all_entries_dfs.append(entries_df)
-
-                if len(all_entries_dfs) == 0:
[…]
