### 2024-06-26 266e5e2
Adjusted and improved the trajectories plot
