+++ glupredkit/plots/trajectories.py
@@ -1,7 +1,10 @@
 from .base_plot import BasePlot
 import ast
 import random
+import re
+import pandas as pd
 import numpy as np
+import matplotlib.colors as mcolors
 import matplotlib.pyplot as plt
 from glupredkit.helpers.unit_config_manager import unit_config_manager
 from datetime import datetime
@@ -11,10 +14,11 @@ class Plot(BasePlot):
     def __init__(self):
         super().__init__()
 
-    def __call__(self, dfs, *args):
+    def __call__(self, dfs, start_date=None, n_samples=12*12, interval=30, alpha=0.2, *args):
         """
         This plot plots predicted trajectories from the measured values. A random subsample of around 24 hours will
         be plotted.
+        # TODO: Add indications of carbs, boluses and exercise?
         """
         def on_hover(event):
             if event.inaxes == ax:
@@ -23,7 +27,7 @@ class Plot(BasePlot):
                     if contains:
                         hover_line.set_alpha(1.0)
                     else:
-                        hover_line.set_alpha(0.2)
+                        hover_line.set_alpha(alpha)
                 fig.canvas.draw_idle()
 
         if unit_config_manager.use_mgdl:
@@ -31,23 +35,53 @@ class Plot(BasePlot):
         else:
             unit = "mmol/L"
 
-        for df in dfs:
-            fig, ax = plt.subplots()
+        # TODO: Check if the dataset is the same for all models, raise error if not
 
+        #for df in dfs:
+        fig, ax = plt.subplots()
+
+        dates = dfs[0]['x_test_dates'][0]
+        timestamp_strings = re.findall(r"Timestamp\('([^']+)'", dates)  # Extract timestamp strings using regex
+        timestamps = [pd.Timestamp(ts).tz_convert(None) for ts in timestamp_strings]
+
+        start_index = 0
+        if start_date:
+            start_index = next((i for i, ts in enumerate(timestamps) if ts >= pd.to_datetime(start_date,
+                                                                                             dayfirst=True)), None)
+        else:
+            # Get random subsample
+            if len(timestamps) - n_samples > 0:
+                start_index = random.randint(0, len(timestamps) - n_samples)
+
+        timestamps = timestamps[start_index:start_index + n_samples]
+
+        y_true = dfs[0][f'target_5'][0]
+        y_true = ast.literal_eval(y_true)
+        y_true = np.array(y_true)[start_index:start_index + n_samples]
+
+        # Use correct unit
+        if unit_config_manager.use_mgdl:
+            ax.axhspan(70, 180, facecolor='blue', alpha=0.2)
+        else:
+            y_true = [unit_config_manager.convert_value(val) for val in y_true]
+            ax.axhspan(unit_config_manager.convert_value(70), unit_config_manager.convert_value(180),
+                       facecolor='blue', alpha=0.2)
+
+        ax.set_title('Blood glucose predicted trajectories')
+        ax.set_xlabel('Datetime')
+        ax.set_ylabel(f'Blood glucose [{unit}]')
+        ax.set_xlim(left=timestamps[0] - pd.Timedelta(hours=0.5), right=timestamps[-1] + pd.Timedelta(hours=0.5))
+        ax.scatter(timestamps, y_true, label='Blood glucose measurements', color='black')
+
+        lines = []
+
+        for df in dfs:
             model_name = df['Model Name'][0]
+            # TODO: Add config file name to the labels
+            # config_file_name = df['config_file_name']
             ph = int(df['prediction_horizon'][0])
 
             prediction_horizons = range(5, ph + 1, 5)
-
-            y_true = df[f'target_5'][0]
-            y_true = ast.literal_eval(y_true)
-
-            n_samples = 12*24
-
[…]
