### 2024-06-26 a078b52
Added a loopalgorithm using the c-embedded swift implementation for the loopalgorithm
