+++ glupredkit/models/loop_algorithm.py
@@ -1,0 +1,200 @@
+from glupredkit.models.base_model import BaseModel
+from glupredkit.helpers.scikit_learn import process_data
+import ctypes
+import pandas as pd
+import json
+
+class Model(BaseModel):
+
+    def __init__(self, prediction_horizon):
+        super().__init__(prediction_horizon)
+        self.basal = None
+        self.insulin_sensitivity_factor = None
+        self.carb_ratio = None
+
+    def _fit_model(self, x_train, y_train, basal=0.75, insulin_sensitivity_factor=66.6, carb_ratio=9, *args):
+        self.basal = basal
+        self.insulin_sensitivity_factor = insulin_sensitivity_factor
+        self.carb_ratio = carb_ratio
+
+        return self
+
+    def _predict_model(self, x_test):
+        predictions = []
+        n_predictions = self.prediction_horizon // 5
+
+        # Load the dynamic Swift / C LoopAlgorithms library
+        swift_lib = ctypes.CDLL('./libLoopAlgorithmToPython.dylib')
+
+        # Specify the argument types and return type of the Swift function
+        swift_lib.generatePrediction.argtypes = [ctypes.c_char_p]
+        swift_lib.generatePrediction.restype = ctypes.POINTER(ctypes.c_double)
+
+        count = 1
+        for idx, row in x_test.iterrows():
+            json_data = self.get_json_from_df(row, idx)
+            json_bytes = json_data.encode('utf-8')  # Convert JSON string to bytes
+            predictions += [self.get_predictions_from_json(swift_lib, json_bytes, n_predictions)]
+
+            print(f'Prediction {count} of {x_test.shape[0]}')
+            count += 1
+
+        return predictions
+
+    def process_data(self, df, model_config_manager, real_time):
+        return process_data(df, model_config_manager, real_time)
+
+    def best_params(self):
+        # Return the best parameters found by GridSearchCV
+        return None
+
+    def get_json_from_df(self, row, start_test_date):
+
+        # Dataframe blood glucose
+        data = {"date": [start_test_date], "value": [row['CGM']]}
+        df_glucose = pd.DataFrame(data)
+        keep_running = True
+        count = 1
+        while keep_running:
+            if 'CGM_' + str(count * 5) in row.index:
+                new_date = start_test_date - pd.Timedelta(minutes=count * 5)
+                new_value = row['CGM_' + str(count * 5)]
+                new_row = pd.DataFrame({'date': [new_date], 'value': [new_value]})
+                df_glucose = pd.concat([df_glucose, new_row], ignore_index=True)
+                count += 1
+            else:
+                keep_running = False
+        df_glucose['date'] = df_glucose['date'].dt.strftime('%Y-%m-%dT%H:%M:%SZ')
+        df_glucose['value'] = df_glucose['value'].astype(int)
+        df_glucose.sort_values(by='date', inplace=True, ascending=True)
+
+        # Dataframe carbohydrates
+        if row['carbs'] > 0:
+            data = {"date": [start_test_date], "grams": [row['carbs']]}
+            df_carbs = pd.DataFrame(data)
+        else:
+            df_carbs = pd.DataFrame()
+        keep_running = True
+        count = 1
+        while keep_running:
+            keep_running = False
+            if 'carbs_' + str(count * 5) in row.index:
+                new_date = start_test_date - pd.Timedelta(minutes=count * 5)
+                new_value = row['carbs_' + str(count * 5)]
+                new_row = pd.DataFrame({'date': [new_date], 'grams': [new_value]})
+                df_carbs = pd.concat([df_carbs, new_row], ignore_index=True)
+                keep_running = True
[…]
