+++ glupredkit/parsers/tidepool.py
@@ -4,6 +4,7 @@ and return the data in a format that can be used as input to the blood glucose p
 """
 from tidepool_data_science_project.makedata.tidepool_api import TidepoolAPI
 import datetime
+import json
 import pandas as pd
 from dateutil import parser
 from .base_parser import BaseParser
@@ -51,7 +52,7 @@ class Parser(BaseParser):
             df_basal = pd.json_normalize(basal_data)[
                 ['time', 'rate']]
             df_basal.time = pd.to_datetime(df_basal.time, errors='coerce')
-            df_basal.rename(columns={"time": "date", "rate": "basal_rate"}, inplace=True)
+            df_basal.rename(columns={"time": "date", "rate": "basal"}, inplace=True)
             df_basal.sort_values(by='date', inplace=True, ascending=True)
             df_basal.set_index('date', inplace=True)
 
@@ -72,22 +73,25 @@ class Parser(BaseParser):
 
             # Resampling all datatypes into the same time-grid
             df = df_glucose.copy()
-            df = df.resample('5T', label='right').mean()
+            df = df.resample('5min', label='right').mean()
 
-            df_carbs = df_carbs.resample('5T', label='right').sum().fillna(value=0)
+            df_carbs = df_carbs.resample('5min', label='right').sum().fillna(value=0)
             df = pd.merge(df, df_carbs, on="date", how='outer')
             df['carbs'] = df['carbs'].fillna(value=0.0)
 
-            df_bolus = df_bolus.resample('5T', label='right').sum()
+            df_bolus = df_bolus.resample('5min', label='right').sum()
             df = pd.merge(df, df_bolus, on="date", how='outer')
+            df['bolus'] = df['insulin']
 
-            df_basal = df_basal.resample('5T', label='right').last()
-            df_basal['basal_rate'] = df_basal['basal_rate'] / 60 * 5  # From U/hr to U (5-minutes)
+            # TODO: The accuracy of the basal delivery could be improved by using duration and ensuring the exact
+            # TODO: sum of the basal delivery is used...
+            df_basal = df_basal.resample('5min', label='right').last()
+            df_basal['basal'] = df_basal['basal'] / 60 * 5  # From U/hr to U (5-minutes)
             df = pd.merge(df, df_basal, on="date", how='outer')
-            df['basal_rate'] = df['basal_rate'].ffill(limit=12 * 24 * 2)
-            df[['insulin', 'basal_rate']] = df[['insulin', 'basal_rate']].fillna(value=0.0)
-            df['insulin'] = df['insulin'] + df['basal_rate']
-            df.drop(columns=(["basal_rate"]), inplace=True)
+            df['basal'] = df['basal'].ffill(limit=12 * 24 * 2)
+            df[['insulin', 'basal']] = df[['insulin', 'basal']].fillna(value=0.0)
+            df['insulin'] = df['insulin'] + df['basal']
+            #df.drop(columns=(["basal_rate"]), inplace=True)
 
             # Add hour of day
             df['hour'] = df.index.hour
@@ -101,6 +105,8 @@ class Parser(BaseParser):
             # Get the current datetime in UTC, given the calendar on current computer
             current_timezone = datetime.datetime.now(datetime.timezone.utc).astimezone().tzinfo
             df.index = df.index.tz_convert(current_timezone)
+
+            df['id'] = 1
 
             return df
 
@@ -140,7 +146,8 @@ class Parser(BaseParser):
 
     def add_activity_states(self, start_date, duration, workout_name, df):
[…]
