### 2024-07-01 #6852 @teubert
> 👋 @teubert, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

My apologies for missing this review request- I was out on paternity leave. I'm happy to see that you were able to find a reviewer.
