### 2024-07-02 #6904 @panagiotisanagnostou
Good evening, everybody. I started working on the review of the GluPredKit!

I this process I stumbled on a small issue which is actually important for the continuation of the review process.

The problem is that I do not possess data to analyze with the package. I need to observe its functionality before I reach a final decision.

Is there a way to obtain data for this process easily? If not real-world data, synthetic ones.

This process, a working example, should also be done as a requirement for the journal's instructions, which at the time are not present in repository.
