### 2024-07-02 #6904 @miriamkw
Thank you @panagiotisanagnostou for pointing that out, and I apologize for the oversight.

I have attached a synthetic dataset which should work. Please note that the parsing-commands will not work without some real data or api keys, so you can instead skip that step and place the synthetic data in the data/raw/ folder.

[synthetic_data_with_datetime.csv](https://github.com/user-attachments/files/16071026/synthetic_data_with_datetime.csv)
