### 2024-07-02 #6904 @miriamkw
I have attached a synthetic dataset. Please place it in the `data/raw/` folder. Note that parsing commands won't work without real data or API keys, so you can skip that step. The plots may look strange due to the use of random numbers. Unfortunately, I cannot provide real data due to its sensitivity.

Please let me know if you run into any other issues.

[synthetic_data_with_datetime.csv](https://github.com/user-attachments/files/16073046/synthetic_data_with_datetime.csv)
