### 2024-07-02 #6904 @danielskatz
@miriamkw - as a track editor observing this review, I suggest that you make this something that anyone can do - post the dataset somewhere public, and update your README and/or docs to tell people where it is and how to test the software with it.  This is just a suggestion, but I think it would make the software easier for people to access and try.
