### 2024-07-02 miriamkw/GluPredKit#64 @miriamkw
Synthetic dataset
Suggested from a JOSS track editor: I suggest that you make this something that anyone can do - post the dataset somewhere public, and update your README and/or docs to tell people where it is and how to test the software with it. This is just a suggestion, but I think it would make the software easier for people to access and try.

Originally posted by @danielskatz in https://github.com/openjournals/joss-reviews/issues/6904#issuecomment-2204093807

Address this issue by adding a synthetic dataset with simulated, realistic data
