### 2024-07-12 843ae79
Working on the loop algorithm using swift code enwrapped in c
