### 2024-07-16 miriamkw/GluPredKit#66 @JonnyTran
Errors on sklearn preprocessor with categorical features
After generating the configurations to use a categorical feature and running `glupredkit train_model random_forest synthetic_configs` give an error:

```bash
Starting pipeline to train model random_forest with configurations in synthetic_configs...
Preprocessing data using standardscaler from file data/raw/synthetic_data_with_datetime.csv, with a prediction horizon of 5 minutes...
Traceback (most recent call last):
  File "/micromamba/envs/glupredkit/bin/glupredkit", line 8, in <module>
    sys.exit(cli())
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/click/core.py", line 1157, in __call__
    return self.main(*args, **kwargs)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/click/core.py", line 1078, in main
    rv = self.invoke(ctx)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/click/core.py", line 1688, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/click/core.py", line 1434, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/click/core.py", line 783, in invoke
    return __callback(*args, **kwargs)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/cli.py", line 207, in train_model
    train_data, _ = helpers.get_preprocessed_data(prediction_horizon, model_config_manager)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/helpers/cli.py", line 106, in get_preprocessed_data
    train_data, test_data = chosen_preprocessor(data)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/preprocessors/standardscaler.py", line 14, in __call__
    train_df, test_df = self.preprocess(df)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/preprocessors/standardscaler.py", line 82, in preprocess
    processed_test_df = self.transform_with_encoder(processed_test_df, encoder)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/preprocessors/standardscaler.py", line 87, in transform_with_encoder
    encoded_cols = encoder.transform(df.loc[:, self.categorical_features])
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/sklearn/utils/_set_output.py", line 313, in wrapped
    data_to_wrap = f(self, X, *args, **kwargs)
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/sklearn/preprocessing/_encoders.py", line 1024, in transform
    X_int, X_mask = self._transform(
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/sklearn/preprocessing/_encoders.py", line 214, in _transform
    raise ValueError(msg)
ValueError: Found unknown categories [np.True_] in column 0 during transform
```

The configuration file was generated with:
> glupredkit generate_config --file-name synthetic_configs --data synthetic_data_with_datetime.csv --prediction-horizon 5 --num-lagged-features 12 --num-features CGM,insulin,carbs --cat-features is_test --preprocessor standardscaler

```json
{
    "data": "synthetic_data_with_datetime.csv",
    "subject_ids": [],
    "preprocessor": "standardscaler",
    "prediction_horizon": 5,
    "num_lagged_features": 12,
    "num_features": [
        "CGM",
        "insulin",
        "carbs"
    ],
    "cat_features": [
        "is_test"
    ],
    "what_if_features": []
}
```

This happens for either the basic or the standardscaler preprocessor.

JOSS review: https://github.com/openjournals/joss-reviews/issues/6904
