### 2024-07-16 #6904 @JonnyTran
Hi everyone. I managed to complete my review checklist and running through the code snippets in the repo README. Here are some of my feedback and suggestions:

On the checklist, I found the `Functionality documentation` was incomplete, specifically because some arguments for certain CLI functions doesn't have the full documentation. For example the `--preprocessor` argument for `glupredkit generate_config` lists two allowed values, basic and standardscaler, that doesn't have description on each value.

I was able to run the code snippets with the provided `synthetic_data_with_datetime.csv` up until the Generate Evaluation Reports and the Draw Plots steps, and I was blocked by the following error:

> glupredkit generate_evaluation_pdf --results-file random_forest__synthetic_configs__5.csv
Generating evaluation report...
Traceback (most recent call last):
  File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/helpers/generate_report.py", line 739, in draw_scatter_plot
    y_test = df[f'target_{ph}'][0]
KeyError: 'target_0'
