### 2024-07-18 #6904 @panagiotisanagnostou
@miriamkw I have seen some issues that need to be addressed.

What I did? I took the entire readme file as have it in your repository and I followed the instructions included in it, to do some dummy predictions for the dataset you provided us.

The reason: We can together fix the errors in the package.

- [ ] Installation of the package in a conda environment. `pip install glupredkit`. OK
- [ ] `glupredkit parse`: OK (I cant check it if you want do a recheck)
- [ ] glupredkit generate_config:
   - [ ] missing argument in the documentation `--what_if_feature`
- [ ] glupredkit train_model:
  - [ ] we have an application that uses several machine learning models there is a need for a better description and or better references on the model used.
    - [ ] At least 1 link from the reference links you have included was dead when I tried to access it, please make sure you address it.
  - [ ] Models:
     - [ ] `double_lstm`: somehow I manage to execute it, by chance I imagine. Othewise the error it produces is: ``ImportError: `keras.optimizers.legacy` is not supported in Keras 3. When using `tf.keras`, to continue using a `tf.keras.optimizers.legacy` optimizer, you can install the `tf_keras` package (Keras 2) and set the environment variable `TF_USE_LEGACY_KERAS=True` to configure TensorFlow to use `tf_keras` when accessing `tf.keras`.``
      - [ ] `loop`: check activated. `ValueError: The input DataFrame is missing the following required columns: basal, bolus` Please check it again.
      - [ ] `lstm`: similarly to double_lstm model somehow I manage to execute it, by chance I imagine. Othewise the error it produces is: ``ImportError: `keras.optimizers.legacy` is not supported in Keras 3. When using `tf.keras`, to continue using a `tf.keras.optimizers.legacy` optimizer, you can install the `tf_keras` package (Keras 2) and set the environment variable `TF_USE_LEGACY_KERAS=True` to configure TensorFlow to use `tf_keras` when accessing `tf.keras`.``
      - [ ] `mtl`: I didn't manage to execute the model. The two errors I saw are
```
File "/home/panagiotis/miniconda3/envs/joss/lib/python3.11/site-packages/glupredkit/cli.py", line 230, in train_model
  model_instance = chosen_model.fit(x_train, y_train)
                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
File "/home/panagiotis/miniconda3/envs/joss/lib/python3.11/site-packages/glupredkit/models/base_model.py", line 14, in fit
  return self._fit_model(x_train, y_train, *args, **kwargs)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
File "/home/panagiotis/miniconda3/envs/joss/lib/python3.11/site-packages/glupredkit/models/mtl.py", line 71, in _fit_model
  opt = tf.keras.optimizers.legacy.Adam(learning_rate=0.00053)
        ^^^^^^^^^^^^^^^^^^^
ImportError: Keras cannot be imported. Check that it is installed.
```
```
ValueError: Could not interpret optimizer identifier: <tf_keras.src.optimizers.legacy.adam.Adam object at 0x7f9363179350>
```
 -  -
      - `naive_linear_regressor`: OK
      - `random_forest`: OK
      - `ridge`: OK
      - [ ] `stacked_plsr`:  I didn't manage to execute the model. The two errors I saw are:
        - `ImportError: Keras cannot be imported. Check that it is installed.`
        - `ValueError: Could not interpret optimizer identifier: <tf_keras.src.optimizers.legacy.adam.Adam object at 0x7f03a410fed0>`
      - `stl`: OK
      - `svr`: OK
      - `tcn`: OK
      - `uva_padova`: (`pip install py_replay_bg`) cant install `py_replay_bg`
      - `zero_order`: OK
- `glupredkit test_model`: OK
  - [ ] !!!interesting bug: imidietly after the training of a lstm model the test_model module is executed correctly. after restart of the machine it produses the following error ValueError: Unrecognized keyword arguments passed to LSTM: {'time_major': False}
- `glupredkit generate_comparison_pdf`: OK
- `glupredkit draw_plots`: OK
  - [ ] missing argument in the documentation `--prediction-horizons`

I know that this massage is extensive but it is necessary. I think there is a need for some work to be done especially because of the cuda interaction.

If you need any details on my system please ask me!
