### 2024-07-19 #6904 @miriamkw
Thank you so much both @JonnyTran and @panagiotisanagnostou for your constructive feedback and help with improving the software. I will look into it as soon as possible and get back to you. I really appreciate the feedback!
