### 2024-08-06 4e0d71e
#65 added flexibility towards file path inputs and enhanced documentation
