### 2024-08-06 miriamkw/GluPredKit#65 @miriamkw
Good point, thank you!

I solved this in the following way:
- Added examples to the "help" statements in the CLI
- Made the commands flexible towards included paths (pro with this solution is that people don't have to repeat their command if they do it "wrong", but the con is that users might add the full path unneccessarily)
