### 2024-08-06 #6904 @miriamkw
Just a small sign of life from my side here; We have been working towards a deadline for the upcoming Friday so we'll pick up this work after that. This morning I got started slightly with addressing an issue from JonnyTran - the progress can be followed in the [`joss_review` branch](https://github.com/miriamkw/GluPredKit/tree/joss_review).
