### 2024-08-13 miriamkw/GluPredKit#66 @miriamkw
Thank you for pointing this out! The "is_test" column is used to separate train and test data in the preprocessing, so the column should not be used as input feature.

I addressed this issue by adding a more descriptive error message when adding this column as an input feature (and the same goes for id):

```
if common_elements:
    raise ValueError(f"'id' and 'is_test' should not be used as a features because they are used as columns to "
                     f"separate subjects and train and test data. Please remove these features from the "
                     f"configuration.")

```

I also added a descriptive error message to whenever a feature defined in the configuration is not present in the input data:
```
if missing_features:
        raise ValueError(f"The following features are defined in the configuration, but are missing from the data: "
                         f"{', '.join(missing_features)}. ")
```

In the synthetic dataset I provided, no column should be used as a categorical feature. However, as part of the JOSS review process, I will integrate a synthetic dataset into the software for all users that want to test the software. In this dataset, I will add categorical features. I will let you know when this is provided
