### 2024-08-13 3e35766
#66 Updated generate config readme and added descriptive error message when config and input data is not aligned
