### 2024-08-13 fbca5a0
Added cgm as obligatory feature and automatically add if missing (with warning)
