### 2024-08-13 471e994
Added --prediction horizons to documentation in draw plots
