### 2024-08-13 72f4fb8
Updated requirements for tensorflow for stability across differenc os
