+++ requirements.txt
@@ -12,7 +12,7 @@ dill
 python-nightscout
 error-grids
 # Optional heavy dependencies
-# tensorflow
+# tensorflow==2.14.0
 # keras-tcn
 # torch
 # opsb-pyloopkit==0.1.0 # This is the original pyloopkit with linear meal model
