### 2024-08-13 c1f00c2
fixed the short prediction horizon error when generating report
