### 2024-08-13 #6904 @miriamkw
@panagiotisanagnostou your feedback was very clear, thank you. I have addressed several of your concerns (although not yet tested properly and merged).

I don’t fully understand: «(pip install py_replay_bg) cant install py_replay_bg». I was not able to reproduce this error. It makes me wonder if something is wrong with our build scripts, as I expected it to be installed with the pypi package.

Could the problem be due to the python version? Unfortunately, this package requires python version < 3.10. The authors have discussed this issue because we don’t want to restrict our whole package based on this single non-fundamental dependency. People might use the package and only need for example PyTorch. I will update the README to recommend python 3.8 or 3.9 (also because of Tensorflow), but please let me know if you have advice on better practices on how to handle this.

If you could please provide me with your:
- Operating system
- Pyton version
- Tensorflow version
- Keras version

that would be very helpful when we conduct further testing!
