### 2024-08-13 #6904 @miriamkw
> I was able to run the code snippets with the provided `synthetic_data_with_datetime.csv` up until the Generate Evaluation Reports and the Draw Plots steps, and I was blocked by the following error:
>
> > glupredkit generate_evaluation_pdf --results-file random_forest__synthetic_configs__5.csv
> > Generating evaluation report...
> > Traceback (most recent call last):
> > File "/micromamba/envs/glupredkit/lib/python3.9/site-packages/glupredkit/helpers/generate_report.py", line 739, in draw_scatter_plot
> > y_test = df[f'target_{ph}'][0]
> > KeyError: 'target_0'
>
> @miriamkw, do you have any suggestion on what I could be doing differently to avoid this error?

Thank you for making me aware of that!
