### 2024-08-17 5732526
Adding readme and code supporting adding synthetic data for experimentation
