### 2024-08-18 27c1aef
added categorical features to example synthetic data
