### 2024-08-18 11f008f
added explanation to recommeneded python versions
