### 2024-08-19 #6904 @panagiotisanagnostou
> @panagiotisanagnostou your feedback was very clear, thank you. I have addressed several of your concerns (although not yet tested properly and merged).
>
> I don’t fully understand: «(pip install py_replay_bg) cant install py_replay_bg». I was not able to reproduce this error. It makes me wonder if something is wrong with our build scripts, as I expected it to be installed with the pypi package.
>
> Could the problem be due to the python version? Unfortunately, this package requires python version < 3.10. The authors have discussed this issue because we don’t want to restrict our whole package based on this single non-fundamental dependency. People might use the package and only need for example PyTorch. I will update the README to recommend python 3.8 or 3.9 (also because of Tensorflow), but please let me know if you have advice on better practices on how to handle this.
>
> If you could please provide me with your:
>
>     * Operating system
>
>     * Pyton version
>
>     * Tensorflow version
>
>     * Keras version
>
>
> that would be very helpful when we conduct further testing

The error occurs at Python `v3.9 ... v3.11` with Ubuntu OS. The error occurs independently of the `tensorflow` or `keras` packages. To duplicate it I suggest you do the execution of the package in a new clean environment after the execution of a simple pip install. The command I command i used is `glupredkit train_model uva_padova my_config --n-steps 1000`, and here is the traceback:

```
Starting pipeline to train model uva_padova with configurations in my_config_1...
Traceback (most recent call last):
  File "/home/panagiotis/miniconda3/envs/joss310/bin/glupredkit", line 8, in <module>
    sys.exit(cli())
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/click/core.py", line 1157, in __call__
    return self.main(*args, **kwargs)
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/click/core.py", line 1078, in main
    rv = self.invoke(ctx)
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/click/core.py", line 1688, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/click/core.py", line 1434, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/click/core.py", line 783, in invoke
    return __callback(*args, **kwargs)
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/glupredkit/cli.py", line 203, in train_model
    model_module = helpers.get_model_module(model)
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/glupredkit/helpers/cli.py", line 55, in get_model_module
    model_module = import lib.import_module(f'glupredkit.models.{model}')
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 1050, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1027, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1006, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 688, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 883, in exec_module
  File "<frozen importlib._bootstrap>", line 241, in _call_with_frames_removed
  File "/home/panagiotis/miniconda3/envs/joss310/lib/python3.10/site-packages/glupredkit/models/uva_padova.py", line 2, in <module>
    from py_replay_bg.py_replay_bg import ReplayBG
ModuleNotFoundError: No module named 'py_replay_bg'
```

As I can see from the package you don't focus on one dependency as you said, but from what I was able to deduce that not all the packages used are very well maintained, that issue on its own may prove to be important. Case and point above from a quick search I did I could find sufficient information about the `py_replay_bg` package.
