+++ requirements.txt
@@ -5,6 +5,7 @@ seaborn
 pandas
 scikit-learn
 tidepool-data-science-project
+reportlab
 numpy
 aiohttp
 click
