### 2024-08-28 d8eae4a
Fixing build error because of output format in evaluate model
