+++ README.md
@@ -182,7 +182,7 @@ glupredkit generate_config
 
 #### Examples 
 
-Example using the synthetic dataset and only the required inputs:
+Example using the synthetic dataset:
 ```
 glupredkit generate_config --file-name my_config_1 --data synthetic_data.csv --prediction-horizon 60 --num-lagged-features 12 --num-features CGM,insulin,carbs --cat-features hour
 ```
