### 2024-08-28 20e59d9
Added CD workflows for testing and upload to Pypi
