+++ .github/workflows/test_release.yml
@@ -1,0 +1,32 @@
+name: Test PyPi Release
+
+on:
+  workflow_dispatch:  # Allows manual trigger
+
+jobs:
+  release:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.9'
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install setuptools wheel twine bumpversion
+
+      - name: Bump version
+        run: bumpversion patch  # Adjust this if you want to bump major or minor versions
+
+      - name: Build and upload package to Test PyPI
+        run: |
+          python setup.py sdist bdist_wheel
+          twine upload --repository testpypi dist/*
+        env:
+          TWINE_USERNAME: ${{ secrets.TEST_PYPI_USERNAME }}
+          TWINE_PASSWORD: ${{ secrets.TEST_PYPI_PASSWORD }}
