### 2024-08-28 2797458
Added bumpversion config file
