+++ .github/workflows/release.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - main  # Trigger only on pushes to the main branch
-  pull_request: # Workaround to get the action to show up before merge to main
+  #pull_request: # Workaround to get the action to show up before merge to main
 
 jobs:
   release:
