### 2024-08-28 71c97f0
Adding git config authentication in CD script
