+++ .github/workflows/test_release.yml
@@ -21,6 +21,11 @@ jobs:
           python -m pip install --upgrade pip
           pip install setuptools wheel twine bumpversion
 
+      - name: Set up Git user
+        run: |
+          git config --local user.name "github-actions"
+          git config --local user.email "github-actions@github.com"
+
       - name: Bump version
         run: bumpversion patch # Adjust this if you want to bump major or minor versions
 
