### 2024-08-28 6c3c4e8
Fixed the main release script to mirror the test script that worked
