### 2024-08-29 4757b16
Fixed the setup directories for synthetic data when using pypi
