### 2024-08-29 6d2e38a
Adjusting bump version script to hopefully work
