+++ .github/workflows/test_release.yml
@@ -7,6 +7,7 @@ on:
 jobs:
   release:
     runs-on: ubuntu-latest
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v3
@@ -19,19 +20,30 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools wheel twine bumpversion
+          pip install setuptools wheel twine bump2version
 
       - name: Set up Git user
         run: |
           git config --local user.name "github-actions"
           git config --local user.email "github-actions@github.com"
 
+      - name: Determine next version
+        id: next_version
+        run: |
+          # Get the current version from setup.py or another file
+          CURRENT_VERSION=$(python setup.py --version)
+          # Calculate the next version (you can adjust this logic if needed)
+          echo "CURRENT_VERSION=${CURRENT_VERSION}" >> $GITHUB_ENV
+          NEXT_VERSION=$(bump2version --dry-run --list patch | grep new_version | cut -d '=' -f 2 | xargs)
+          echo "NEXT_VERSION=${NEXT_VERSION}" >> $GITHUB_ENV
+          echo "Next version: ${NEXT_VERSION}"
+
       - name: Bump version
         id: bump_version
         run: |
-          bumpversion patch # Adjust this if you want to bump major or minor versions
+          bump2version patch  # Change to minor or major if needed
           git add .  # Add updated files to the commit
-          git commit -m "Bump version to $(bumpversion --list | grep current_version | cut -d '=' -f 2 | xargs)"
+          git commit -m "Bump version to $(python setup.py --version)"
           git tag $(python setup.py --version)  # Create a new tag with the new version
           git push
           git push --tags  # Push the new tag to the remote repository
@@ -47,3 +59,4 @@ jobs:
         env:
           TWINE_USERNAME: ${{ secrets.TEST_PYPI_USERNAME }}
           TWINE_PASSWORD: ${{ secrets.TEST_PYPI_PASSWORD }}
+
