+++ .bumpversion.cfg
@@ -4,3 +4,5 @@ commit = True
 tag = True
 
 [bumpversion:file:setup.py]
+
+[bumpversion:file:pyproject.toml]
