+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
 current_version = 1.0.1
-commit = True
+commit = False
 tag = True
 
 [bumpversion:file:setup.py]
