+++ .github/workflows/test_release.yml
@@ -35,24 +35,12 @@ jobs:
       - name: Bump version
         id: bump_version
         run: |
-          # Bump the version in setup.py and pyproject.toml
-          bump2version patch --allow-dirty --verbose  # Ensure you see what's happening
-      
-          # Capture the new version
-          NEW_VERSION=$(bump2version --dry-run --list patch | grep new_version | cut -d '=' -f 2 | xargs)
-          echo "New version: ${NEW_VERSION}"  # Debug output
-      
-          # Ensure changes are staged for commit
-          git add setup.py pyproject.toml
-      
-          # Check if there are any changes to commit
-          if git diff-index --quiet HEAD --; then
-            echo "No changes to commit."
-          else
-            git commit -m "Bump version to ${NEW_VERSION}"
-            git push origin HEAD
-            git push --tags  # Push tags if needed
-          fi
+          bump2version patch  # Automatically bumps, commits, and tags
+
+      - name: Push changes
+        run: |
+          git push 
+          git push --tags
 
       - name: Clean up old builds
         run: |
