+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
-current_version = 1.0.2
-commit = False
+current_version = 1.0.3
+commit = True
 tag = True
 
 [bumpversion:file:setup.py]
