+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "glupredkit"
-version = "1.0.2"
+version = "1.0.3"
 authors = [
   { name="Miriam K. Wolff", email="miriamkwolff@outlook.com" },
 ]
