### 2024-08-29 e3b5b93
Updated the main script to bump versions and push changes as well
