+++ .github/workflows/release.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools wheel twine bumpversion
+          pip install setuptools wheel twine bump2version
 
       - name: Set up Git user
         run: |
@@ -28,7 +28,16 @@ jobs:
           git config --local user.email "github-actions@github.com"
 
       - name: Bump version
-        run: bumpversion patch  # Adjust this if you want to bump major or minor versions
+        run: bump2version patch  # Adjust this if you want to bump major or minor versions
+
+      - name: Push changes
+        run: |
+          git push
+          git push --tags
+
+      - name: Clean up old builds
+        run: |
+          rm -rf dist build *.egg-info
 
       - name: Build and upload package to PyPI
         run: |
