### 2024-08-29 849edef
Bump version: 1.0.4 → 1.0.5
