### 2024-08-29 7db795f
Bump version: 1.0.5 → 1.0.6
