+++ glupredkit/cli.py
@@ -166,6 +166,8 @@ def parse(parser, username, password, start_date, file_path, end_date, output_fi
 def generate_config(file_name, data, subject_ids, preprocessor, prediction_horizon, num_lagged_features, num_features,
                     cat_features, what_if_features):
 
+    print("JUST TESTING THAT PRINTS SHOW UP!")
+
     generate_model_configuration(file_name, data, subject_ids, preprocessor, int(prediction_horizon),
                                  int(num_lagged_features), num_features, cat_features, what_if_features)
     click.echo(f"Storing configuration file to data/configurations/{file_name}...")
