### 2024-08-29 89eac9f
Bump version: 1.0.6 → 1.0.7
