### 2024-08-29 5dfa5e3
Fixed the saving synthetic data
