### 2024-08-29 8379428
Bump version: 1.0.7 → 1.0.8
