### 2024-08-29 4e7c3d3
Removed duplicate check
