+++ glupredkit/helpers/model_config_manager.py
@@ -10,12 +10,6 @@ import warnings
 
 def generate_model_configuration(file_name, data, subject_ids, preprocessor, prediction_horizon, num_lagged_features,
                                  num_features, cat_features, what_if_features):
-
-    # Check if the 'data' string is a valid file in the 'data/raw/' folder
-    data_path = os.path.join('data/raw', data + '.csv')
-    if not os.path.isfile(data_path):
-        raise ValueError(f"Data file '{data + '.csv'}' not found in 'data/raw/' folder.")
-
     if not "CGM" in num_features:
         warnings.warn("Warning: The 'CGM' column is required for numerical features. Please note that CGM is "
                       "added to your configuration by the system.", UserWarning)
