+++ requirements.txt
@@ -1,7 +1,7 @@
-matplotlib
+matplotlib>=2.1.0
 reportlab
 svglib
-seaborn
+seaborn>=0.9.0
 pandas
 scikit-learn
 tidepool-data-science-project
