### 2024-08-30 69eb419
Fix short prediction horizons, and removing temporary workflow changes
