### 2024-08-30 9f57158
Joss review (#67)
* #65 added flexibility towards file path inputs and enhanced documentation

* #65 Fixed the test_model build error due to  prefix ambigouity with pytests

* #66 Updated generate config readme and added descriptive error message when config and input data is not aligned

* Added cgm as obligatory feature and automatically add if missing (with warning)

* Added --prediction horizons to documentation in draw plots

* Updated requirements for tensorflow for stability across differenc os

* Added more descriptive error message for loop model requried columns

* Added description 
[…]
