+++ tests/test_cli_helper.py
@@ -1,5 +1,5 @@
 import pytest
-from glupredkit.helpers.cli import (split_string, validate_file_name, validate_subject_ids, validate_prediction_horizon,
+from glupredkit.helpers.cli import (split_string, validate_config_file_name, validate_subject_ids, validate_prediction_horizon,
                                     validate_num_lagged_features, validate_feature_list, validate_test_size)
 from click import BadParameter
 
@@ -10,7 +10,7 @@ def test_split_string():
     assert split_string(None) == []
 
 
-def test_validate_file_name():
+def test_validate_config_file_name():
     # Mock objects for ctx and param, which are not used in the function
     mock_ctx = None
     mock_param = None
@@ -19,17 +19,18 @@ def test_validate_file_name():
     test_cases = [
         ("filename.txt", "filename"),
         ("filename", "filename"),
-        ("filename.with.multiple.dots.txt", "filename"),
-        ("filename.with.multiple.dots", "filename"),
+        ("filename.with.multiple.dots.txt", "filename.with.multiple.dots"),
+        ("filename.with.multiple.dots", "filename.with.multiple"),
+        ("filename/with/path/filename.json", "filename"),
         ("filename.", "filename"),
-        (".hiddenfile", ""),
+        (".hiddenfile", ".hiddenfile"),
         ("", ""),
         (123, "123"),  # Non-string input
         (None, "None")  # None as a string
     ]
 
     for input_value, expected_output in test_cases:
-        assert validate_file_name(mock_ctx, mock_param, input_value) == expected_output, f"Failed on input {input_value}"
+        assert validate_config_file_name(mock_ctx, mock_param, input_value) == expected_output, f"Failed on input {input_value}"
 
 
 # Mock objects for ctx and param, which are not used in the functions
