+++ glupredkit/cli.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 import click
 import dill
+import requests
+import warnings
 import os
 import ast
 import importlib
@@ -141,10 +143,11 @@ def parse(parser, username, password, start_date, file_path, end_date, output_fi
 
 
 @click.command()
-@click.option('--file-name', help='Name of the configuration file (without file extension).',
-              callback=helpers.validate_file_name, required=True)
-@click.option('--data', help='Name of the data file from data/raw/.',
-              callback=helpers.validate_file_name, required=True)
+@click.option('--file-name', help='Give a file name to the configuration file (e.g., "my_config").',
+              callback=helpers.validate_config_file_name, required=True)
+@click.option('--data', help='Name of the input data file (e.g., "data.csv"). '
+                             'This file should be located in data/raw/.',
+              callback=helpers.check_if_data_file_exists, required=True)
 @click.option('--subject-ids', help='The ids you want to include in the model training and testing. '
                                     ' will include all of the subjects.', callback=helpers.validate_subject_ids,
               required=False, default='')
@@ -157,13 +160,27 @@ def parse(parser, username, password, start_date, file_path, end_date, output_fi
 @click.option('--num-lagged-features', help='Number of lagged features.',
               callback=helpers.validate_num_lagged_features, required=True)
 @click.option('--num-features', help='Comma-separated list of numerical features.',
-              callback=helpers.validate_feature_list, required=True)
+              callback=helpers.validate_feature_list, required=True, default='CGM')
 @click.option('--cat-features', help='Comma-separated list of categorical features.',
               callback=helpers.validate_feature_list, required=False, default='')
 @click.option('--what-if-features', help='Comma-separated list of categorical features.',
               callback=helpers.validate_feature_list, required=False, default='')
 def generate_config(file_name, data, subject_ids, preprocessor, prediction_horizon, num_lagged_features, num_features,
                     cat_features, what_if_features):
+
+    if data == 'synthetic_data':
+        cwd = os.getcwd()
+        url = 'https://raw.githubusercontent.com/miriamkw/GluPredKit/main/example_data/synthetic_data.csv'
+        save_folder = 'data/raw/'
+        save_path = os.path.join(cwd, save_folder, 'synthetic_data.csv')
+
+        if not os.path.exists(save_path):
+            response = requests.get(url)
+            with open(save_path, 'wb') as file:
+                file.write(response.content)
+
+            click.echo(f"Synthetic data saved to {save_path}")
+
     generate_model_configuration(file_name, data, subject_ids, preprocessor, int(prediction_horizon),
                                  int(num_lagged_features), num_features, cat_features, what_if_features)
     click.echo(f"Storing configuration file to data/configurations/{file_name}...")
@@ -196,6 +213,12 @@ def train_model(model, config_file_name, epochs, n_cross_val_samples, n_steps):
     1) Process data using the given configurations
     2) Train the given models for the given prediction horizons in the configuration
     """
+    # Filtering out UserWarning because we are using an old Keras file format on purpose
+    warnings.filterwarnings(
+        "ignore",
[…]
