+++ glupredkit/helpers/model_config_manager.py
@@ -5,14 +5,15 @@ models before preprocessing and training.
 
 import json
 import os
+import warnings
 
 
 def generate_model_configuration(file_name, data, subject_ids, preprocessor, prediction_horizon, num_lagged_features,
                                  num_features, cat_features, what_if_features):
-    # Check if the 'data' string is a valid file in the 'data/raw/' folder
-    data_path = os.path.join('data/raw', data + '.csv')
-    if not os.path.isfile(data_path):
-        raise ValueError(f"Data file '{data + '.csv'}' not found in 'data/raw/' folder.")
+    if not "CGM" in num_features:
+        warnings.warn("Warning: The 'CGM' column is required for numerical features. Please note that CGM is "
+                      "added to your configuration by the system.", UserWarning)
+        num_features += ["CGM"]
 
     # Check if 'preprocessor' is a valid preprocessor module
     preprocessor_module = f'glupredkit.preprocessors.{preprocessor}'
