### 2024-09-04 0e31f71
Modified the python versions in the tests
