+++ .github/workflows/test_cli.yml
@@ -8,15 +8,15 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.8", "3.9"]
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install libopenmpi-dev
-        run: sudo apt-get update && sudo apt-get install -y libopenmpi-dev
+      #- name: Install libopenmpi-dev
+      #  run: sudo apt-get update && sudo apt-get install -y libopenmpi-dev
       - name: Install test dependencies
         run: |
           python -m pip install --upgrade pip
