### 2024-09-04 756f1dd
Updated command docs for optional dependencies
