### 2024-09-04 3f3f498
Removed unneccessary install in test process
