+++ .github/workflows/test_cli.yml
@@ -15,8 +15,6 @@ jobs:
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
-      #- name: Install libopenmpi-dev
-      #  run: sudo apt-get update && sudo apt-get install -y libopenmpi-dev
       - name: Install test dependencies
         run: |
           python -m pip install --upgrade pip
