### 2024-09-04 b7f84a1
Simplify linting
