+++ .github/workflows/test_cli_helper.yml
@@ -22,12 +22,7 @@ jobs:
           pip install flake8 pytest
       - name: Lint code with flake8
         run: |
-          # stop the build if there are Python syntax errors or undefined names
-          flake8 glupredkit --count --select=E9,F63,F7,F82 --show-source --statistics
-          flake8 tests --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          flake8 glupredkit --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-          flake8 tests --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+          flake8 glupredkit tests --count --select=E9,F63,F7,F82 --show-source --statistics --max-complexity=10 --max-line-length=127 --exit-zero
       - name: Run Integration Tests
         run: |
           pytest tests/test_cli_helper.py
