### 2024-09-04 02f2430
Testing new solution for build error windows cli
