+++ .github/workflows/test_cli.yml
@@ -9,7 +9,8 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ["3.9", "3.10", "3.11"]
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        #os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [windows-latest]
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python ${{ matrix.python-version }}
@@ -21,7 +22,6 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt     
           pip install flake8 pytest
-          pip install .[heavy]  # Install the package with [heavy] dependencies
       - name: Lint code with flake8
         run: |
           flake8 glupredkit tests --count --select=E9,F63,F7,F82 --show-source --statistics --max-complexity=10 --max-line-length=127 --exit-zero
