### 2024-09-04 a206bb5
Added more file path handling
