### 2024-09-05 4726bc7
Try to fix broken model config manager test
