+++ tests/test_cli.py
@@ -179,7 +179,6 @@ def test_evaluate_model(runner, temp_dir):
         # Check if the model test file was created
         output_file_name = f'{model}__{config}__60.csv'
         output_path = Path('data') / 'tested_models' / output_file_name
-        #assert os.path.exists(os.path.join(output_path, output_file_name))
         assert output_path.exists(), f"Expected file {output_path} was not created"
 
 
@@ -196,7 +195,6 @@ def test_generate_evaluation_pdf(runner, temp_dir):
         # Check if reports were generated
         output_file_name = f'{model}__{config}__60.pdf'
         output_path = Path('data') / 'reports' / output_file_name
-        #assert os.path.exists(os.path.join(output_path, output_file_name))
         assert output_path.exists(), f"Expected file {output_path} was not created"
 
 
