### 2024-09-05 409c951
Trying to test tensorflow models in integration test
