+++ tests/test_cli.py
@@ -136,7 +136,7 @@ def test_train_model(runner, temp_dir):
     args_list = [
         # ['blstm', config_file_name, '--epochs', epochs],
         # ['loop', config_file_name, '--n-cross-val-samples', 10],
-        # ['lstm', config_file_name, '--epochs', epochs],
+        ['lstm', config_file_name, '--epochs', epochs],
         # ['mtl', config_file_name, '--epochs', epochs],
         ['naive_linear_regressor', config_file_name],
         # ['random_forest', config_file_name],
@@ -161,7 +161,6 @@ def test_train_model(runner, temp_dir):
         # Check if the model file was created
         output_file_name = f'{args[0]}__{config_file_name}__60.pkl'
         output_path = Path('data') / 'trained_models' / output_file_name
-        #assert os.path.exists(os.path.join(output_path, output_file_name))
         assert output_path.exists(), f"Expected file {output_path} was not created"
 
 
@@ -169,7 +168,7 @@ def test_evaluate_model(runner, temp_dir):
     runner = CliRunner()
 
     config = 'my_config_1'
-    models = ['naive_linear_regressor', 'ridge', 'zero_order']
+    models = ['lstm', 'naive_linear_regressor', 'ridge', 'zero_order']
 
     for model in models:
 
@@ -186,7 +185,7 @@ def test_generate_evaluation_pdf(runner, temp_dir):
     runner = CliRunner()
 
     config = 'my_config_1'
-    models = ['naive_linear_regressor', 'ridge', 'zero_order']
+    models = ['lstm', 'naive_linear_regressor', 'ridge', 'zero_order']
 
     for model in models:
         result = runner.invoke(generate_evaluation_pdf, ['--results-file', f'{model}__{config}__60.csv'])
