+++ .github/workflows/test_unit_config_manager.yml
@@ -1,5 +1,6 @@
 name: test_unit_config_manager
-on: push
+on: #push
+  workflow_dispatch:
 permissions:
   contents: read
 jobs:
