+++ .github/workflows/test_metrics.yml
@@ -1,5 +1,6 @@
 name: test_metrics
-on: push
+on: #push
+  workflow_dispatch:
 permissions:
   contents: read
 jobs:
