+++ .github/workflows/test_models.yml
@@ -1,5 +1,6 @@
 name: test_models
-on: push
+on: #push
+  workflow_dispatch:
 permissions:
   contents: read
 jobs:
