+++ .github/workflows/test_cli_helper.yml
@@ -1,5 +1,6 @@
 name: test_cli_helper
-on: push
+on: #push
+  workflow_dispatch:
 permissions:
   contents: read
 jobs:
