+++ .github/workflows/test_cli.yml
@@ -20,6 +20,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt     
+          pip install .[heavy]  # Install the heavy dependencies
           pip install flake8 pytest
       - name: Lint code with flake8
         run: |
