### 2024-09-05 8de43be
Add innstall mpi dependencies for mac/ubuntu
