+++ .github/workflows/test_cli.yml
@@ -16,6 +16,15 @@ jobs:
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Install MPI dependencies (Ubuntu/Mac)
+        if: runner.os == 'Linux' || runner.os == 'macOS'
+        run: |
+          if [ "$(uname)" == "Linux" ]; then
+            sudo apt-get update
+            sudo apt-get install -y libopenmpi-dev
+          elif [ "$(uname)" == "Darwin" ]; then
+            brew install mpich
+          fi
       - name: Install test dependencies
         run: |
           python -m pip install --upgrade pip
