### 2024-09-05 02565d0
Shortening the test_cli test again
