+++ tests/test_cli.py
@@ -135,8 +135,8 @@ def test_train_model(runner, temp_dir):
 
     args_list = [
         # ['blstm', config_file_name, '--epochs', epochs],
-        ['loop', config_file_name, '--n-cross-val-samples', 10],
-        ['lstm', config_file_name, '--epochs', epochs],
+        # ['loop', config_file_name, '--n-cross-val-samples', 10],
+        # ['lstm', config_file_name, '--epochs', epochs],
         # ['mtl', config_file_name, '--epochs', epochs],
         ['naive_linear_regressor', config_file_name],
         # ['random_forest', config_file_name],
@@ -185,7 +185,7 @@ def test_generate_evaluation_pdf(runner, temp_dir):
     runner = CliRunner()
 
     config = 'my_config_1'
-    models = ['loop', 'lstm', 'naive_linear_regressor', 'ridge', 'zero_order']
+    models = ['naive_linear_regressor', 'ridge', 'zero_order']
 
     for model in models:
         result = runner.invoke(generate_evaluation_pdf, ['--results-file', f'{model}__{config}__60.csv'])
@@ -208,7 +208,7 @@ def test_draw_plots(runner, temp_dir):
     runner = CliRunner()
 
     config = 'my_config_1'
-    results_files = f'naive_linear_regressor__{config}__60.csv,ridge__{config}__60.csv,'
+    results_files = f'naive_linear_regressor__{config}__60.csv,ridge__{config}__60.csv'
 
     result = runner.invoke(draw_plots, ['--results-files', results_files, '--plots', 'scatter_plot', '--prediction-horizons', '30'])
     assert result.exit_code == 0
