+++ .github/workflows/test_cli.yml
@@ -16,20 +16,20 @@ jobs:
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install MPI dependencies (Ubuntu/Mac)
-        if: runner.os == 'Linux' || runner.os == 'macOS'
-        run: |
-          if [ "$(uname)" == "Linux" ]; then
-            sudo apt-get update
-            sudo apt-get install -y libopenmpi-dev
-          elif [ "$(uname)" == "Darwin" ]; then
-            brew install mpich
-          fi
+      #- name: Install MPI dependencies (Ubuntu/Mac)
+      #  if: runner.os == 'Linux' || runner.os == 'macOS'
+      #  run: |
+      #    if [ "$(uname)" == "Linux" ]; then
+      #      sudo apt-get update
+      #      sudo apt-get install -y libopenmpi-dev
+      #    elif [ "$(uname)" == "Darwin" ]; then
+      #      brew install mpich
+      #    fi
       - name: Install test dependencies
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt     
-          pip install .[heavy]  # Install the heavy dependencies
+      #    pip install .[heavy]  # Install the heavy dependencies
           pip install flake8 pytest
       - name: Lint code with flake8
         run: |
