+++ .github/workflows/test_cli.yml
@@ -8,8 +8,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: [ "3.9", "3.10", "3.11" ]
+        os: [ ubuntu-latest, macos-latest, windows-latest ]
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python ${{ matrix.python-version }}
@@ -30,7 +30,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt     
-      #    pip install .[heavy]  # Install the heavy dependencies
+          #    pip install .[heavy]  # Install the heavy dependencies
           pip install flake8 pytest
       - name: Lint code with flake8
         run: |
