### 2024-09-05 45b9087
Fix hanging up plot test
