### 2024-09-05 fda2415
Reset test settings
