+++ .github/workflows/test_model_config_manager.yml
@@ -1,6 +1,5 @@
 name: test_model_config_manager
-on: #push
-  workflow_dispatch:
+on: push
 permissions:
   contents: read
 jobs:
