+++ .github/workflows/test_models.yml
@@ -1,6 +1,5 @@
 name: test_models
-on: #push
-  workflow_dispatch:
+on: push
 permissions:
   contents: read
 jobs:
