+++ .github/workflows/test_metrics.yml
@@ -1,6 +1,5 @@
 name: test_metrics
-on: #push
-  workflow_dispatch:
+on: push
 permissions:
   contents: read
 jobs:
