+++ .github/workflows/test_cli_helper.yml
@@ -1,6 +1,5 @@
 name: test_cli_helper
-on: #push
-  workflow_dispatch:
+on: push
 permissions:
   contents: read
 jobs:
