### 2024-09-05 386d081
Test enhancements (#68)
* Modified the python versions in the tests

* Updated command docs for optional dependencies

* Changed the requirement for python version because of dependency constraints

* Removed unneccessary install in test process

* Simplify linting

* Running tests on different oses

* Testing new solution for build error windows cli

* Testing path error fix windows tests

* Testing a fix in save models for windows support

* Added more file path handling

* Refactoring to using Path

* Try to fix broken model config manager test

* Hopefully the last path fix

* Another wind
[…]
