### 2024-09-05 2edbec8
Bump version: 1.0.11 → 1.0.12
