### 2024-09-07 #6904 @miriamkw
Dear all, we have finally addressed the feedback and performed testing of the changes. PyPi as well as the main branch in GitHub is up to date. Thank you so much for the helpful conversation in the process.

Below we have written a systematic overview over all the comments raised along with our response and implementations.

**Synthetic Dataset**
Danielskatz suggested that we incorporated a synthetic dataset into the software open for everyone:
- README is updated with instructions for how to use it
    - When using the generate_config command, the user can add the argument --data synthetic_data.csv. The synthetic dataset will be copied into the «data/raw» folder.
- We added a bolus and basal column so that the dataset can support all the models (the physiological models require those columns)
- We added a categorical feature column («hour») so that it can be tested (addressing https://github.com/miriamkw/GluPredKit/issues/65)
- The dataset is stored in «example_data/synthetic_data.csv»

**[Review Panagiotisanagnostou](https://github.com/openjournals/joss-reviews/issues/6904#issuecomment-2237620672):**
* CLI Arguments: Missing CLI arguments have been added to the documentation.
* Keras and TensorFlow Issues:
    * We have fixed the Keras-related issues by enforcing TensorFlow version 2.14.0, which ensures compatibility with the correct Keras version. It’s important that dependencies are installed using pip install glupredkit[hard] to avoid these problems.
    * The test_model bug you encountered might have been due to version mismatches in TensorFlow and Keras. While we haven't reproduced this bug in our testing, we will monitor it closely.
    * The [hard] option was introduced to separate heavy dependencies, which are now optional. This prevents unnecessary installation for users who don’t need them.
    * We are confident that we have resolved these problems, as we were able to recreate the errors and no longer observe them. However, we would greatly appreciate it if you could test the software again, given that some of the issues were peculiar.
* Documentation and Installation:
    * We have updated the installation instructions in the README.
    * Other documentation issues you identified have also been resolved.

Regarding the concerns about certain packages being poorly maintained, we agree that this is a limitation. However, we believe that those dependencies, despite being less maintained, are essential for certain advanced research applications. They are used for physiological modeling, one of them based on an FDA approved simulator, and can hence serve as a benchmark against which data-driven models can be validated, ensuring they align with physiological expectations. To balance these considerations, we implemented optional installation of heavy dependencies via the [heavy] option.

We want to add that your feedback was instrumental in helping us address deeper issues with our version constraints, dependencies, and build scripts. We’ve updated the required Python version to >= 3.9 and expanded our integration tests to cover all supported Python versions and operating systems (ubuntu-latest, macos-latest, and windows-latest).

So again, thank you so much for your thorough review and for clearly explaining the errors.

**Review JonnyTran:**
We would like to refer you to the issues you created earlier in this thread, as we have addressed them in our responses there. Additionally, we want to express our gratitude for providing such clear and detailed descriptions when raising these issues.

- Issue: https://github.com/openjournals/joss-reviews/issues/6904#issuecomment-2231987704
    - Response/fix: https://github.com/openjournals/joss-reviews/issues/6904#issuecomment-2287097001
- Issue: https://github.com/miriamkw/GluPredKit/issues/66
    - Response and commits in the comments of the issue
- Issue: [https://github.com/miriamkw/GluPredKit/issues/65](https://github.com/miriamkw/GluPredKit/issues/66)
    - Response and commits in the comments of the issue

**Additional improvements:**
- Added a CD workflow for automatic version bumping and PyPi upload when pushing new code to the main branch. This ensures that PyPi and the main branch is always up to date with each other.
- In general, while performing testing, we changed the documentation where we found that the descriptions could be clearer

**Testing**
Kind reminder that if you want to test the changes, use a clean python 3.9 environment, use pip install **glupredkit[heavy]**, and make sure to install the most recent version (v1.0.12).

If you encounter an mpi4py issue due to the py_replay_bg dependency, please refer to the "System-Wide Dependencies for MPI" section in the documentation. Unfortunately, this fix cannot be included in the PyPi build distribution.

**Pull requests**
- [Main changes JOSS Review](https://github.com/miriamkw/GluPredKit/pull/67)
- [Test enhancements](https://github.com/miriamkw/GluPredKit/pull/68)

We apologize for the time it took; this process has been a valuable learning experience. However, we believe it will contribute to a more stable development of the software moving forward!
