### 2024-09-16 #6904 @jromanowska
@panagiotisanagnostou - were you able to review the updates?

@JonnyTran - do you have more comments that the authors should address? If not, could you tick off the last point [on your checklist](https://github.com/openjournals/joss-reviews/issues/6904#issuecomment-2197989997)?
