### 2024-09-19 #6904 @panagiotisanagnostou
I apologize for the delayed response.

It appears that most of my previous comments have been addressed. However, due to the recent updates in the package, I have some concerns that may need to be resolved.

The primary issue involves the `py_replay_bg` package. Despite the additional instructions provided in the README.md, I was unable to install it. The installation process fails during the compilation of the `mpi4py` package.

The second concern is related to the compatibility between TensorFlow and Python versions. Although it is explicitly stated on the PyPI package installation page that the package supports Python version 3.7 and newer, this does not seem to be the case. The issue appears to stem from a version conflict between TensorFlow and Python.

@jromanowska, my checklist is now complete, and I believe that once these final issues are addressed or commented upon, I will be satisfied.
