### 2024-09-19 #6904 @miriamkw
>
> The primary issue involves the `py_replay_bg` package. Despite the additional instructions provided in the README.md, I was unable to install it. The installation process fails during the compilation of the `mpi4py` package.

Regarding the installation issue with `py_replay_bg` and `mpi4py`, we recognize that macOS and Ubuntu systems require a manual installation of `mpi4py` before running `pip install`. We included instructions in the README to address this ([see the section on system-wide dependencies](https://github.com/miriamkw/GluPredKit?tab=readme-ov-file#system-wide-dependencies-for-mpi)), but we can add a reference next to the `pip install glupredkit[heavy]` command to ensure this is clear.

We explored automating this process through the PyPI build, but after multiple attempts, it seems that it may not be feasible.

>
> The second concern is related to the compatibility between TensorFlow and Python versions. Although it is explicitly stated on the PyPI package installation page that the package supports Python version 3.7 and newer, this does not seem to be the case. The issue appears to stem from a version conflict between TensorFlow and Python.

As for the Python version conflict with TensorFlow, we agree with this issue, which led us to update the required Python version to 3.9 and newer in the revised release ([v1.0.12](https://pypi.org/project/GluPredKit/)). We understand that this introduces stricter version requirements, but we believe it will provide greater stability and help avoid errors for users. I’ve double-checked both the documentation and source code, and the version constraint appears to be updated throughout.

>
> @jromanowska, my checklist is now complete, and I believe that once these final issues are addressed or commented upon, I will be satisfied.

Thank you again! Please let us know if any further clarification is needed.
