### 2024-09-20 #6904 @panagiotisanagnostou
@miriamkw Thank you for your immediate response. The only comment left from me, is that I think you need to include the [hard] version to the documentation. I think this a very well structured application for the execution of simple predictions.

I am very happy to be reviewer to an application like this

@jromanowska With this final comment I conclude my review. Thank you for the gait coordination.
