### 2024-09-23 #6904 @jromanowska
@miriamkw - there is a wrongly formatted reference in line 65, please check that, correct paper.md and re-generate the pdf with `@editorialbot generate pdf`
