### 2024-09-23 41bae58
Enhanced the trajectories plot
