### 2024-09-23 #6904 @miriamkw
I have now updated the paper fixing the wrongly formatted reference and added ORCIDs for all authors.

The documentation in the newest version of GluPredKit contains documentation of the [heavy] option for dependencies, as requested in the final comment by [panagiotisanagnostou](https://github.com/panagiotisanagnostou).

GluPredKit v1.0.12 is archived in Zenodo with the following DOI: https://doi.org/10.5281/zenodo.13828438
I confirm that I have made sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper, and that the license listed for the archive is the same as the software license.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13828438.svg)](https://doi.org/10.5281/zenodo.13828438)

@jromanowska that should cover the steps in _Author Tasks_.
