### 2024-09-26 #6904 @miriamkw
> @miriamkw as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks. Please see the open point below that needs your attention:
>
>
> ## Checks on archive
> * [ ]  Archive listed title and authors matches paper
>   -> Please remove the text `miriamkw/GluPredKit: ` from the archive title, so that the title fully matches the paper title.

@Kevin-Mattheus-Moerman thank you. I have now fixed the title on the archive: https://zenodo.org/records/13828438.
